upgrade aws provider version to ~>6.0 (#9)
With the latest AWS Provider release for Terraform, a new feature called
Enhanced Region Support is now available. This allows users to deploy
resources across multiple regions by specifying the region argument
directly within resource and data blocks, rather than maintaining
separate provider configurations for each region.

This allows our configuration to be made simpler as we can rely on a
single provider setup instead of having to configure multiple providers
with one in each region as we have done previously.

The goal of this PR is to upgrade the Terraform version used in this
module and enable users to pass a region argument, which will
automatically cascade through the module’s resources.